latest master from upstream

+++ extensions/histogram/src/test/java/io/druid/query/aggregation/histogram/ApproximateHistogramErrorBenchmark.java
@@ -1,20 +1,18 @@
 /*
  * Druid - a distributed column store.
- * Copyright (C) 2012, 2013  Metamarkets Group Inc.
+ * Copyright 2012 - 2015 Metamarkets Group Inc.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package io.druid.query.aggregation.histogram;
@@ -152,26 +150,26 @@ public class ApproximateHistogramErrorBenchmark
     float err1 = 0;
     float err2 = 0;
     for (int j = 0; j < hcounts.length; j++) {
-      err1 += Math.abs((hcounts[j] - ahcounts1[j]) / numValues);
-      err2 += Math.abs((hcounts[j] - ahcounts2[j]) / numValues);
+      err1 += (float)Math.abs((hcounts[j] - ahcounts1[j]) / numValues);
+      err2 += (float)Math.abs((hcounts[j] - ahcounts2[j]) / numValues);
     }
 
     if (debug) {
       float sum = 0;
       for (double v : hcounts) {
-        sum += v;
+        sum += (float)v;
       }
       System.out.println("Exact Histogram Sum:");
       System.out.println(sum);
       sum = 0;
       for (double v : ahcounts1) {
-        sum += v;
+        sum += (float)v;
       }
       System.out.println("Approximate Histogram Sum:");
       System.out.println(sum);
       sum = 0;
       for (double v : ahcounts2) {
-        sum += v;
+        sum += (float)v;
       }
       System.out.println("Approximate Histogram Rule Fold Sum:");
       System.out.println(sum);
